Add NixieAnimation sketch
This sketch was used to develop a pleasing animation for changing nixie
tube numbers under normal operation. The end result is an animation that
changes nixies one at a time from left to right, i.e. most significant
to least significant. The digits in each nixie tube are stacked, not in
numerical order, but in one that's meant to prevent visual obstruction.
For each tube, we identify where in the stack the current digit resides,
and where in the stack the target digit resides. We then "move towards"
the target digit by lighting up the digit in the stack that's one step
closer to the target digit relative to the current digit. This creates
an effect wherein numbers appear to recede into or come out of the tube,
relative to the viewer.