Hotfix DP5 lighting up when button is LOW
Long story short, I've looked at the schematics of the Doayee board and
at the datasheets for the HV5122, and using the board with the ESP8266
seems to be doable without level shifting. I tried the TXB0108 shifter,
but I had a ton of noise due to clock issues, so I'm connecting the 5v
nixie board to the 3.3v ESP8266 directly. I'm not the first:

https://old.reddit.com/r/arduino/comments/83s09b/ive_always_wanted_a_nixie_clock_and_finally_made/

Everything seems very stable, except this one thing.

When the button is held down, if there is a dot in the second-to-last
segment (currentDots[3]), then the dot in the last segment lights up
as well. We hide it here by blanking the digits and lighting up all
the dots during a state transition.

Looking at the schematics of the nixie board, DP5 is unlike the other
dots: due to running out of pins on the HV5122's, it's controlled by
a subcircuit with a 74LVC1G79 and a FET. This subcircuit is connected
directly to 5v, GND, and CLK, with data coming from the second HV5122.
I don't understand fully how it works, but I'm guessing that the data
signal is at 3.3v, which is normally just enough to get an accurate
binary signal to that subcircuit, but when the button is pushed, that
signal voltage decreases just enough to cause it to float. Or something
to that effect. I have little understanding of what's actually going
wrong, but I doubt it fixable without modifying the board. I don't want
to revisit level shifting at this stage in the project unless there's
no other option.